+++ osmchadjango/changeset/tests/test_changeset_views.py
@@ -1,10 +1,13 @@
 import json
+import mock
 
 from django.contrib.gis.geos import Polygon
 from django.core.urlresolvers import reverse
+from django.test import TestCase, override_settings
 
 from social_django.models import UserSocialAuth
 from rest_framework.test import APITestCase
+import oauth2 as oauth
 
 from ...users.models import User
 from ...feature.tests.modelfactories import FeatureFactory
@@ -161,6 +164,11 @@ class TestChangesetFilteredViews(APITestCase):
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data['count'], 2)
+        self.assertTrue(
+            response.data['features'][0]['properties']['check_user'].startswith(
+                'user '
+                )
+            )
 
     def test_unchecked_changesets_view(self):
         url = reverse('changeset:unchecked-list')
@@ -575,6 +583,13 @@ class TestCheckChangesetViews(APITestCase):
             user=self.user,
             provider='openstreetmap',
             uid='123123',
+            extra_data={
+                'id': '123123',
+                'access_token': {
+                    'oauth_token': 'aaaa',
+                    'oauth_token_secret': 'bbbb'
+                    }
+                }
             )
         self.tag_1 = TagFactory(name='Illegal import')
         self.tag_2 = TagFactory(name='Vandalism')
@@ -643,12 +658,16 @@ class TestCheckChangesetViews(APITestCase):
         self.assertIsNone(self.changeset_2.check_user)
         self.assertIsNone(self.changeset_2.check_date)
 
-    def test_set_harmful_changeset_put(self):
+    @override_settings(ENABLE_POST_CHANGESET_COMMENTS=True)
+    @mock.patch.object(oauth.Client, 'request')
+    def test_set_harmful_changeset_put(self, mock_oauth_client):
         """User can set a changeset of another user as harmful with a PUT request.
         We can also set the tags of the changeset sending it as data.
         """
+        mock_oauth_client.return_value = [{'status': '200'}]
         self.client.login(username=self.user.username, password='password')
         data = {'tags': [self.tag_1.id, self.tag_2.id]}
+        # ENABLE POST CHANGESET COMMENTS for this test
         response = self.client.put(
             reverse('changeset:set-harmful', args=[self.changeset_2.pk]),
             data
@@ -669,6 +688,13 @@ class TestCheckChangesetViews(APITestCase):
             self.tag_2,
             self.changeset_2.tags.all()
             )
+        mock_oauth_client.assert_called_with(
+            'https://api.openstreetmap.org/api/0.6/changeset/{}/comment/'.format(
+                self.changeset_2.pk
+                ),
+            method='POST',
+            body=mock.ANY
+        )
 
     def test_set_harmful_changeset_with_invalid_tag_id(self):
         """Return a 400 error if a user try to add a invalid tag id to a changeset.
@@ -719,10 +745,13 @@ class TestCheckChangesetViews(APITestCase):
         self.assertIsNone(self.changeset_2.check_user)
         self.assertIsNone(self.changeset_2.check_date)
 
-    def test_set_good_changeset_put(self):
+    @override_settings(ENABLE_POST_CHANGESET_COMMENTS=True)
+    @mock.patch.object(oauth.Client, 'request')
+    def test_set_good_changeset_put(self, mock_oauth_client):
         """User can set a changeset of another user as good with a PUT request.
         We can also set the tags of the changeset sending it as data.
         """
+        mock_oauth_client.return_value = [{'status': '200'}]
         self.client.login(username=self.user.username, password='password')
         data = {'tags': [self.tag_1.id, self.tag_2.id]}
         response = self.client.put(
@@ -744,6 +773,13 @@ class TestCheckChangesetViews(APITestCase):
             self.tag_2,
             self.changeset_2.tags.all()
             )
+        mock_oauth_client.assert_called_with(
+            'https://api.openstreetmap.org/api/0.6/changeset/{}/comment/'.format(
+                self.changeset_2.pk
+                ),
+            method='POST',
+            body=mock.ANY
+        )
 
     def test_set_good_changeset_with_invalid_tag_id(self):
         """Return a 400 error if a user try to add a invalid tag id to a changeset.
@@ -824,6 +860,13 @@ class TestUncheckChangesetView(APITestCase):
             user=self.user,
             provider='openstreetmap',
             uid='123123',
+            extra_data={
+                'id': '123123',
+                'access_token': {
+                    'oauth_token': 'aaaa',
+                    'oauth_token_secret': 'bbbb'
+                    }
+                }
             )
         self.suspect_changeset = SuspectChangesetFactory()
         self.good_changeset = GoodChangesetFactory(check_user=self.user)
@@ -849,7 +892,10 @@ class TestUncheckChangesetView(APITestCase):
         self.assertEqual(self.harmful_changeset.tags.count(), 1)
         self.assertIn(self.tag, self.harmful_changeset.tags.all())
 
-    def test_uncheck_harmful_changeset(self):
+    @override_settings(ENABLE_POST_CHANGESET_COMMENTS=True)
+    @mock.patch.object(oauth.Client, 'request')
+    def test_uncheck_harmful_changeset(self, mock_oauth_client):
+        mock_oauth_client.return_value = [{'status': '200'}]
         self.client.login(username=self.user.username, password='password')
         response = self.client.put(
             reverse('changeset:uncheck', args=[self.harmful_changeset.pk]),
@@ -861,6 +907,13 @@ class TestUncheckChangesetView(APITestCase):
         self.assertIsNone(self.harmful_changeset.check_user)
         self.assertIsNone(self.harmful_changeset.check_date)
         self.assertEqual(self.harmful_changeset.tags.count(), 1)
+        mock_oauth_client.assert_called_with(
+            'https://api.openstreetmap.org/api/0.6/changeset/{}/comment/'.format(
+                self.harmful_changeset.pk
+                ),
+            method='POST',
+            body=mock.ANY
+        )
 
     def test_uncheck_good_changeset(self):
         self.client.login(username=self.user.username, password='password')